==> 0.create-file-directories.R <==
# SET-UP -----------------------------------------------------------------------
# Load libraries

# CREATE DIRECTORIES -----------------------------------------------------------
# For original datafiles
dir.create("1.data/1.Dataset_order_202507/Results", recursive = T)
dir.create("1.data/1.Dataset_order_202510/Results", recursive = T)

# For finalized data
dir.create("1.data/4.finalized", recursive = T)
